captioning/requirements.txt
# Core ML dependencies
torch>=2.0.0
transformers>=4.50.0
accelerate>=0.20.0

# Data handling
datasets>=2.0.0
huggingface_hub>=0.20.0
pandas>=1.5.0
pyarrow>=12.0.0
Pillow>=9.0.0

# Data download (for download_data.py)
img2dataset>=1.45.0

# Utilities
tqdm>=4.64.0
